+++ views/admin-cms-content-element/_form-announce.php
@@ -1,0 +1,16 @@
+<?php
+/* @var $this yii\web\View */
+/* @var $model \skeeks\cms\models\CmsContentElement */
+/* @var $relatedModel \skeeks\cms\relatedProperties\models\RelatedPropertiesModel */
+?>
+<?= $form->fieldSet(\Yii::t('skeeks/cms','Announcement')); ?>
+    <?= $form->field($model, 'image_id')->widget(
+        \skeeks\cms\widgets\formInputs\StorageImage::className()
+    ); ?>
+    <?= $form->field($model, 'description_short')->widget(
+        \skeeks\cms\widgets\formInputs\comboText\ComboTextInputWidget::className(),
+        [
+            'modelAttributeSaveType' => 'description_short_type',
+        ]);
+    ?>
+<?= $form->fieldSetEnd() ?>