+++ views/admin-cms-content-element/_form-detail.php
@@ -1,0 +1,19 @@
+<?php
+/* @var $this yii\web\View */
+/* @var $model \skeeks\cms\models\CmsContentElement */
+/* @var $relatedModel \skeeks\cms\relatedProperties\models\RelatedPropertiesModel */
+?>
+<?= $form->fieldSet(\Yii::t('skeeks/cms','In detal')); ?>
+
+    <?= $form->field($model, 'image_full_id')->widget(
+        \skeeks\cms\widgets\formInputs\StorageImage::className()
+    ); ?>
+
+    <?= $form->field($model, 'description_full')->widget(
+        \skeeks\cms\widgets\formInputs\comboText\ComboTextInputWidget::className(),
+        [
+            'modelAttributeSaveType' => 'description_full_type',
+        ]);
+    ?>
+
+<?= $form->fieldSetEnd() ?>