+++ views/admin-cms-content-element/_form-images.php
@@ -1,0 +1,16 @@
+<?php
+/* @var $this yii\web\View */
+/* @var $model \skeeks\cms\models\CmsContentElement */
+/* @var $relatedModel \skeeks\cms\relatedProperties\models\RelatedPropertiesModel */
+?>
+<?= $form->fieldSet(\Yii::t('skeeks/cms','Images/Files')); ?>
+
+    <?= $form->field($model, 'images')->widget(
+        \skeeks\cms\widgets\formInputs\ModelStorageFiles::className()
+    ); ?>
+
+    <?= $form->field($model, 'files')->widget(
+        \skeeks\cms\widgets\formInputs\ModelStorageFiles::className()
+    ); ?>
+
+<?= $form->fieldSetEnd()?>